Update orphaned filter logic and add a unit test
While doing this commit I found some issues with the previous patching commit. Fixing them required making a new connections util module to avoid circular dependencies
