+++ src/fides/api/ops/api/v1/endpoints/saas_config_endpoints.py
@@ -14,7 +14,6 @@ from starlette.status import (
 )
 
 from fides.api.ops.api import deps
-from fides.api.ops.api.v1.endpoints.connection_endpoints import validate_secrets
 from fides.api.ops.api.v1.scope_registry import (
     CONNECTION_AUTHORIZE,
     SAAS_CONFIG_CREATE_OR_UPDATE,
@@ -58,6 +57,7 @@ from fides.api.ops.service.connectors.saas.connector_registry_service import (
     upsert_dataset_config_from_template,
 )
 from fides.api.ops.util.api_router import APIRouter
+from fides.api.ops.util.connection_util import validate_secrets
 from fides.api.ops.util.oauth_util import verify_oauth_client
 from fides.lib.exceptions import KeyOrNameAlreadyExists
